pyproject.toml
[project]
name = "lawcode"
version = "0.1.0"
description = "A simple OpenAI wrapper to test legal text generation"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
]

[tool.ruff]
line-length = 100
